Add Autofac modules
