+++ DiscordBot/AutofacModules/MiscellaneousAutofacModule.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Autofac;
+
+namespace DiscordBot.AutofacModules
+{
+    public class MiscellaneousAutofacModule : Module
+    {
+        protected override void Load(ContainerBuilder builder)
+        {
+            builder.RegisterType<Random>().SingleInstance();
+        }
+    }
+}
